Remove nonfunctional tag transation file code
The super constructor call in DefaultDALClient.__init__ was being
invoked incorrectly, and thus wasn't actually running. If it had been
running, it would have been guaranteed to throw an error (by calling the
non-existent method .listdir() on a string). Since it confused me a lot
and doesn't do anything, I've just removed it.

Whether or not IDALClient should exist at all is another question -
looks like the only reason it does is because this code is based on a
Java library, where interfaces are common practice. Left it for now.